Consolidate FormDef tests
